requirements.txt
blinker==1.9.0
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
et_xmlfile==2.0.0
Flask==3.1.2
fonttools==4.61.1
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mysql-connector-python==9.5.0
numpy==2.4.2
openpyxl==3.1.5
packaging==26.0
pillow==12.1.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
six==1.17.0
Werkzeug==3.1.5
